order

+++ database/migrations/2021_02_07_163345_create_cart_products.php
@@ -18,7 +18,7 @@ class CreateCartProducts extends Migration
             $table->bigInteger('product_id');
             $table->bigInteger('user_id')->unsigned();
             $table->string('product_type');
-            $table->decimal('unit_price');
+            $table->integer('unit_price');
             $table->string('quantity');
             $table->timestamps();
